+++ examples/vanilla/CHANGELOG.md
@@ -1,5 +1,11 @@
 # vanilla
 
+## 0.0.2
+
+### Patch Changes
+
+- @eagle-tracker/core@1.1.1
+
 ## 0.0.1
 
 ### Patch Changes