+++ examples/vue3-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # vue3-test
 
+## 0.0.2
+
+### Patch Changes
+
+- @eagle-tracker/vue3@1.1.1
+
 ## 0.0.1
 
 ### Patch Changes