+++ packages/web-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @eagle-tracker/core
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @eagle-tracker/types@2.0.0
+
 ## 1.1.0
 
 ### Minor Changes